+++ CHANGES.md
@@ -31,6 +31,7 @@
   + `Primes.mobius`
   + `Primes.derivative`
 - add `Primes.order`, `Primes.order_with_known_multiple`, `Primes.order_mod_prime_pow`
+- `Primes.factors` now performs some iterations of Fermat’s factor searching
 - BREAKING: rename sequence-related functions:
   + `Arith.sum_seq` -> `sum_of_seq`
   + `Arith.prod_seq` -> `prod_of_seq`